feat(mdx): add markdown.crossCompilerCache for speed up build process
