+++ packages/core/src/node/mdx/loader.ts
@@ -1,37 +1,11 @@
-import path from 'node:path';
-import { createProcessor } from '@mdx-js/mdx';
 import type { Rspack } from '@rsbuild/core';
-import type { FrontMatterMeta, Header, UserConfig } from '@rspress/shared';
 import { logger } from '@rspress/shared/logger';
-import { extractTextAndId, loadFrontMatter } from '@rspress/shared/node-utils';
 
-import type { PluginDriver } from '../PluginDriver';
-import type { RouteService } from '../route/RouteService';
-import {
-  applyReplaceRules,
-  escapeMarkdownHeadingIds,
-  normalizePath,
-} from '../utils';
-import { createMDXOptions } from './options';
-import type { TocItem } from './remarkPlugins/toc';
-
-interface LoaderOptions {
-  config: UserConfig;
-  docDirectory: string;
-  checkDeadLinks: boolean;
-  routeService: RouteService;
-  pluginDriver: PluginDriver;
-}
-
-export interface PageMeta {
-  toc: TocItem[];
-  title: string;
-  headingTitle: string;
-  frontmatter?: FrontMatterMeta;
-}
+import { compile, compileWithCrossCompilerCache } from './processor';
+import type { MdxLoaderOptions } from './types';
 
 export default async function mdxLoader(
-  this: Rspack.LoaderContext<LoaderOptions>,
+  this: Rspack.LoaderContext<MdxLoaderOptions>,
   source: string,
 ) {
   this.cacheable(true);
@@ -42,72 +16,34 @@ export default async function mdxLoader(
   const { config, docDirectory, checkDeadLinks, routeService, pluginDriver } =
     options;
 
-  // Separate frontmatter and content in MDX source
-  const { frontmatter, emptyLinesSource } = loadFrontMatter(
-    source,
-    filepath,
-    docDirectory,
-    true,
-  );
-
-  // For loader error stack, so we use the source with frontmatter @see https://github.com/web-infra-dev/rspress/issues/2010
-  const replacedSource = applyReplaceRules(
-    emptyLinesSource,
-    config.replaceRules,
-  );
-
-  // Support custom id like `#hello world {#custom-id}`.
-  // TODO > issue: `#hello world {#custom-id}` will also be replaced.
-  const preprocessedContent = escapeMarkdownHeadingIds(replacedSource);
+  const crossCompilerCache = config?.markdown?.crossCompilerCache ?? true;
 
   try {
-    let pageMeta: PageMeta = {
-      title: '',
-      toc: [] as TocItem[],
-      headingTitle: '',
-    };
-
-    const frontmatterTitle = extractTextAndId(frontmatter.title)[0];
-
-    const mdxOptions = await createMDXOptions(
-      docDirectory,
-      config,
-      checkDeadLinks,
-      routeService,
-      filepath,
-      pluginDriver,
-    );
-    const compiler = createProcessor(mdxOptions);
-
-    compiler.data('pageMeta' as any, { toc: [], title: '' });
-    const vFile = await compiler.process({
-      value: preprocessedContent,
-      path: filepath,
-    });
-
-    const compileResult = String(vFile);
-    const compilationMeta = compiler.data('pageMeta' as any) as {
-      toc: Header[];
-      title: string;
-    };
-    const headingTitle = extractTextAndId(compilationMeta.title)[0];
-    pageMeta = {
-      ...compilationMeta,
-      title: frontmatterTitle || headingTitle,
-      headingTitle,
-      frontmatter,
-    } as PageMeta;
-
-    const result = `const frontmatter = ${JSON.stringify(frontmatter)};
-${compileResult}
-MDXContent.__RSPRESS_PAGE_META = {};
-
-MDXContent.__RSPRESS_PAGE_META["${encodeURIComponent(
-      normalizePath(path.relative(docDirectory, filepath)),
-    )}"] = ${JSON.stringify(pageMeta)};
-`;
-    callback(null, result);
-  } catch (e: unknown) {
+    // TODO wrong but good enough for now (example: "build --watch")
+    if (crossCompilerCache && process.env.NODE_ENV === 'production') {
+      const compileResult = await compileWithCrossCompilerCache({
+        source,
+        filepath,
+        docDirectory,
+        checkDeadLinks,
+        config,
+        pluginDriver,
+        routeService,
+      });
+      callback(null, compileResult);
+    } else {
+      const compileResult = await compile({
+        source,
+        filepath,
+        docDirectory,
+        checkDeadLinks,
+        config,
+        pluginDriver,
+        routeService,
+      });
+      callback(null, compileResult);
+    }
+  } catch (e) {
     if (e instanceof Error) {
       logger.error(`MDX compile error: ${e.message} in ${filepath}`);
       logger.debug(e);
